# specs/phase-3/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Phase 3 Chatbot API
  description: API for the AI-powered Todo Chatbot
  version: 1.0.0
servers:
  - url: http://localhost:8000
    description: Development server
paths:
  /api/{user_id}/chat:
    post:
      summary: Main chat endpoint for user interactions
      parameters:
        - name: user_id
          in: path
          required: true
          description: The ID of the user making the request
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
                  description: The user's message to the chatbot
                  example: "Add a task to buy groceries"
                conversation_id:
                  type: string
                  description: Optional conversation ID to continue existing conversation
                  example: "123e4567-e89b-12d3-a456-426614174000"
      responses:
        '200':
          description: Successful response from the chatbot
          content:
            application/json:
              schema:
                type: object
                properties:
                  response:
                    type: string
                    description: The chatbot's response
                    example: "I've added the task 'buy groceries' to your list."
                  conversation_id:
                    type: string
                    description: The conversation ID for this interaction
                  task_operations:
                    type: array
                    items:
                      type: object
                      properties:
                        operation:
                          type: string
                          enum: [add, update, complete, delete]
                        task_data:
                          type: object
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...